+++ src/App.js
@@ -1,6 +1,8 @@
 import React, { Component } from "react";
 import styled from "styled-components";
-import Filtros from "./components/Filtros";
+import Filtros from "./components/Filtros/Filtros";
+import Home from "./components/Home/Home"
+import Carrinho from "./components/Carrinho/Carrinho";
 
 const PaginaInicial = styled.div `
 display:flex;
@@ -12,159 +14,6 @@ height: 100%;
 min-height:100vh;
 padding: 30px;
 `
-const ContainerHome = styled.div`
-display:flex;
-justify-content:space-between;
-gap: 10px;
-width: 100%;
-height: 100%;
-min-height:100vh;
-/* padding: 30px; */
-`
-
-const HeaderHome = styled.div `
-display:flex;
-justify-content: space-between;
-align-items: center;
-width: 100%;
-height:5%;
-background-color: blue;
-
-select{
-  width: 100px;
-  height: 20px;
-  margin: 7px;
-}
-`
-const ContainerCard = styled.div `
-display: flex;
-width: 100%;
-background-color: pink;
-flex-wrap: wrap;
-padding: 24px;
-justify-content: center;
-`
-const Card = styled.div `
-border: 1px solid black;
-display:flex;
-flex-direction: column;
-justify-content: space-between;
-align-items: flex-start;
-margin: 0 30px;
-width: 200px;
-height: 250px;
-img{
-  width: 100%;
-  height: 70%;
-}
-`
-const CarrinhoInicial = styled.div `
-border: 1px solid black;
-width: 20%;
-height: 450px;
-overflow: auto;
-display: flex;
-flex-direction: column;
-`
-const CarrinhoItem = styled.div `
-display: flex;
-flex-direction: row;
-justify-content: space-between;
-margin: 2px 0px;
-padding: 0px 10px;
-background-color: red;
-`
-// const ContainerFiltros = styled.div `
-// border: 1px solid black;
-// width: 20%;
-// height: 450px;
-// padding: 16px;
-// `
-
-class Home extends React.Component{
-  state = {
-    // inputValorMinimo: 0,
-    // inputValorMaximo: 1000,
-    // inputBuscaNome: ""
-  }
-
-  // onChangeValorMinimo = (event) => {
-  //   this.setState({inputValorMinimo: event.target.value})
-  // }
-  // onChangeValorMaximo = (event) => {
-  //   this.setState({inputValorMaximo: event.target.value})
-  // }
-  // onChangeBuscaNome = (event) => {
-  //   this.setState({inputBuscaNome: event.target.value})
-  // }
-
-  render(){
-    return (
-    <div>
-      <HeaderHome>
-        <p>Quantidade de produtos:{this.props.produtos.length}</p>
-        <label>Ordenação:</label>
-        <select>
-          <option>Crescente</option>
-          <option>Decrescente</option>
-        </select>
-      </HeaderHome>
-      <CardHome
-        produtos={this.props.produtos}
-        addproduto={this.props.addproduto}
-        minimo={this.props.minimo}
-        maximo={this.props.maximo}
-        busca={this.props.busca}
-      />
-    </div>
-    )
-  };
-};
-class CardHome extends React.Component{
-  render(){
-    const listaProdutos = this.props.produtos.filter((item) => {
-      return item.value >= this.props.minimo
-    }).filter((item) => {
-      return item.value <= this.props.maximo
-    }).filter((item) => {
-      return item.name.toLowerCase().includes(this.props.busca.toLowerCase())
-    }).map((dados)=>{
-      return (
-      <Card>
-        <img src={dados.imageUrl} alt="imagem do: {dados.name}"/>
-        <p>{dados.name}</p>
-        <p>{dados.value}</p>
-        <button onClick={()=>this.props.addproduto(dados.id)}>Adicionar ao Carrinho</button>
-      </Card>
-    )})
-    return(
-      <ContainerCard>
-      {listaProdutos}
-      </ContainerCard>
-    )
-  }
-}
-
-class Carrinho extends React.Component{
-  render(){
-    const renderizaProduto = this.props.carrinho.map((dados)=>{
-      return (
-        <CarrinhoItem> 
-          <p>{dados.id.length}</p>
-          <p>{dados.name}</p>
-          <button onClick={() => this.props.removeProduto(dados.id)}>Remover</button>
-
-        </CarrinhoItem>
-      );
-    })
-    return(
-      <CarrinhoInicial>
-      <h3>Carrinho:</h3>
-      {renderizaProduto}
-    </CarrinhoInicial>
-    )
-  };
-};
 class App extends React.Component{
   state={
     produtos:[{
