+++ src/components/Filtros/Filtro.styled.js
@@ -1,0 +1,8 @@
+import styled from "styled-components";
+
+export const ContainerFiltros = styled.div `
+border: 1px solid black;
+width: 20%;
+height: 450px;
+padding: 16px;
+`
