+++ src/components/Home/Home.styled.js
@@ -1,0 +1,26 @@
+import styled from "styled-components";
+
+// export const ContainerHome = styled.div`
+// display:flex;
+// justify-content:space-between;
+// gap: 10px;
+// width: 100%;
+// height: 100%;
+// min-height:100vh;
+// /* padding: 30px; */
+// `
+
+export const HeaderHome = styled.div `
+display:flex;
+justify-content: space-between;
+align-items: center;
+width: 100%;
+height:5%;
+background-color: blue;
+
+select{
+  width: 100px;
+  height: 20px;
+  margin: 7px;
+}
+`
